+++ src/ray/gcs/gcs_server/test/gcs_actor_manager_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <ray/gcs/test/gcs_test_util.h>
+#include <ray/gcs/gcs_server/test/gcs_server_test_util.h>
 
 #include <memory>
 #include "gtest/gtest.h"
