+++ src/ray/gcs/gcs_server/test/gcs_server_rpc_test.cc
@@ -15,7 +15,7 @@
 #include "gtest/gtest.h"
 #include "ray/common/test_util.h"
 #include "ray/gcs/gcs_server/gcs_server.h"
-#include "ray/gcs/test/gcs_test_util.h"
+#include <ray/gcs/gcs_server/test/gcs_server_test_util.h>
 #include "ray/rpc/gcs_server/gcs_rpc_client.h"
 
 namespace ray {
